Cleveland Processed Data- Project Proposal
I loaded the data and added column headers. Then I only selected the columns which have the variables that we are using for our analysis. I then split the data into a testing and a training set, where 75% of the data is in the training set. I made a table showing the average values for each of the predictor variables for each class of the num column, then plotted those trends using a line graph.