Changes made after the proposal
Combined 1-4 to be "disease" and had 0 be "no disease" so that our sample is now more balanced. Also checked for NA. Still trying to troubleshoot how to make a box plot or add error bars to the bar graphs to show that they are good predictor varibles. Additionally, need to. figure out how to change the y axis of the graphs. I will proceed to build the classifier and. come back to the graphs at a later. time.